Создать тесты для FilmController и UserController.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -23,5 +23,5 @@ public class Film {
     LocalDate releaseDate;
 
     @Positive
-    Long duration;
+    Integer duration;
 }
